Uso de Jinja2/requirements.txt
Flask==2.3.2
redis==4.5.5
python-dotenv==1.0.0
uuid==1.30